+++ tests/e2e/test_kg_builder_pipeline_e2e.py
@@ -151,7 +151,7 @@ async def test_pipeline_builder_happy_path(
     """
     driver.execute_query("MATCH (n) DETACH DELETE n")
     embedder.embed_query.return_value = [1, 2, 3]
-    llm.invoke.side_effect = [
+    llm.ainvoke.side_effect = [
         LLMResponse(
             content="""{
                         "nodes": [
@@ -250,22 +250,24 @@ async def test_pipeline_builder_happy_path(
                 ("Organization", "LED_BY", "Person"),
             ],
         },
+        "extractor": {"document_info": {"path": "my document path"}},
     }
     res = await kg_builder_pipeline.run(pipe_inputs)
     # llm must have been called for each chunk
-    assert llm.invoke.call_count == 3
+    assert llm.ainvoke.call_count == 3
     # result must be success
     assert res == {"writer": {"status": "SUCCESS"}}
     # check component's results
     chunks = kg_builder_pipeline.get_results_for_component("splitter")
     assert len(chunks["chunks"]) == 3
     graph = kg_builder_pipeline.get_results_for_component("extractor")
-    # 3 entities + 3 chunks
+    # 3 entities + 3 chunks + 1 document
     nodes = graph["nodes"]
-    assert len(nodes) == 6
+    assert len(nodes) == 7
     label_counts = dict(Counter([n["label"] for n in nodes]))
     assert label_counts == {
         "Chunk": 3,
+        "Document": 1,
         "Person": 2,
         "Organization": 1,
     }
@@ -273,14 +275,20 @@ async def test_pipeline_builder_happy_path(
     # + 3 rels between entities and their chunk
     # + 2 "NEXT_CHUNK" rels
     relationships = graph["relationships"]
-    assert len(relationships) == 7
+    assert len(relationships) == 10
     type_counts = dict(Counter([r["type"] for r in relationships]))
-    assert type_counts == {"FROM_CHUNK": 3, "KNOWS": 1, "LED_BY": 1, "NEXT_CHUNK": 2}
+    assert type_counts == {
+        "FROM_CHUNK": 3,
+        "FROM_DOCUMENT": 3,
+        "KNOWS": 1,
+        "LED_BY": 1,
+        "NEXT_CHUNK": 2,
+    }
     # then check content of neo4j db
     created_nodes = driver.execute_query("MATCH (n) RETURN n")
-    assert len(created_nodes.records) == 6
+    assert len(created_nodes.records) == 7
     created_rels = driver.execute_query("MATCH ()-[r]->() RETURN r")
-    assert len(created_rels.records) == 7
+    assert len(created_rels.records) == 10
 
     created_chunks = driver.execute_query("MATCH (n:Chunk) RETURN n").records
     assert len(created_chunks) == 3
@@ -305,7 +313,7 @@ async def test_pipeline_builder_failing_chunk_raise(
     """
     driver.execute_query("MATCH (n) DETACH DELETE n")
     embedder.embed_query.return_value = [1, 2, 3]
-    llm.invoke.side_effect = [
+    llm.ainvoke.side_effect = [
         LLMResponse(
             content="""{
                         "nodes": [
@@ -384,7 +392,7 @@ async def test_pipeline_builder_failing_chunk_do_not_raise(
     """
     driver.execute_query("MATCH (n) DETACH DELETE n")
     embedder.embed_query.return_value = [1, 2, 3]
-    llm.invoke.side_effect = [
+    llm.ainvoke.side_effect = [
         LLMResponse(content="invalid json"),
         LLMResponse(
             content="""{
@@ -446,7 +454,7 @@ async def test_pipeline_builder_failing_chunk_do_not_raise(
     ).component.on_error = OnError.IGNORE  # type: ignore[attr-defined, unused-ignore]
     res = await kg_builder_pipeline.run(pipe_inputs)
     # llm must have been called for each chunk
-    assert llm.invoke.call_count == 3
+    assert llm.ainvoke.call_count == 3
     # result must be success
     assert res == {"writer": {"status": "SUCCESS"}}
     # check component's results