+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Next
 
+### Changed
+- When saving the lexical graph in a KG creation pipeline, the document is also saved as a specific node, together with relationships between each chunk and the document they were created from.
+
 ## 0.5.0
 
 ### Fixed